use base class for bucket and bucket type properties
